+++ src/main/php/streams/StreamException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file is part of stubbles.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @package  stubbles
+ */
+namespace stubbles\streams;
+/**
+ * Thrown when reading or writing on a stream fails.
+ */
+class StreamException extends \Exception
+{
+    // intentionally empty
+}